requirements.txt
# Core dependencies
ccxt==4.1.22
pandas==2.1.4
numpy==1.24.3
aiohttp==3.9.1
websockets==12.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pymongo==4.6.0

# ML/AI
scikit-learn==1.3.2
tensorflow==2.15.0
xgboost==2.0.2
ta-lib==0.4.28
pandas-ta==0.3.14b0

# Web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Monitoring & Logging
prometheus-client==0.19.0
loguru==0.7.2
python-telegram-bot==20.7

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
cryptography==41.0.7
schedule==1.2.0
backtrader==1.9.78.123

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0